+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `external_link_to` helper method for links which should open in a new tab/window
+- `Metus.public_controller_class` to configure the public controller to be used by Metus